requirements.txt
Django==3.2.8
django-cors-headers==3.11.0
djangorestframework==3.13.1
json5==0.9.6
mysqlclient==2.1.0
Pillow==8.2.0
PyMySQL==1.0.2

